+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added conversion of audios for sending recorded audio, now it is possible to send mp3 audios and not just ogg
 * Route to fetch all groups that the connection is part of
 * Route to fetch all privacy settings
+* Route to update the privacy settings
 
 ### Fixed